feat: add custom arch support w/ env var
We've recently added a darwin_arm64 build to terraform-demux, but that
can cause issues in Terraform workspaces that require older versions of
Terraform - depending on the version, some don't have darwin_arm64
builds, and even the ones that do may use Terraform providers that don't
have darwin_arm64 builds. In either case, the user is unable to run
terraform-demux as-is, and is forced to install the amd64 version.

This commit allows the user to override the architecture by specifying
an environment variable, and updates the README to match. It is expected
that users will set up the recommended shell alias and invoke the amd64
version of terraform when necessary, but otherwise use the native arm64
builds.